+++ test/support/factory_man_demo/factory/child_factory.ex
@@ -108,7 +108,7 @@ defmodule FactoryManDemo.Factory.ChildFactory do
 
   # ── Factory options ──────────────────────────────────────────────
 
-  deffactory raw_user(params \\ %{}), struct: User, build_params?: false do
+  deffactory raw_user(params \\ %{}), struct: User, body: :struct do
     username = Map.get(params, :username, "raw-user-#{System.os_time()}")
     %User{username: username}
   end